requirements.txt
aiogram
python-dotenv
apscheduler
pandas
openpyxl